BUG: spacings between consecutive digits were not calculated well and made some of them overlap
In previous revisions, we provided an augmentation method
`random_number_char_spacingn` for `PIL.ImageFont`. This augmentation
is targeting to simulate the format of datetime-like string in
Chinese characters. In some format of datetime-like string, the
spacings of digits and characters are not the same, and the former
one is usually narrower than the latter one.

In order to simulate this format of text, the spacings between
consecutive digits can be randomly changed (narrowed down), and it's
actually achieved by subtracting an offset in x-axis. That offset
was calculated based on the font size of previous character, but
it might result in a value larger than current spacing between
characters.

That is, given a normal character spacing width: 8, and the font size
of previous character is 16, the increment of x_offset will be 12
according to this formula:
```
x_offset += int(prev_font_size / 4.0 * 3)
-> x_offset += int(16 / 4.0 * 3)
-> x_offset += 12
```
The increment is obviously greater than the normal character spacing,
so that it resulted in the character overlap.

To solve this issue, we currently fixed the increment of offset to
a unit of normal character spacing. So that it would only make the
new spacing between iterated characters to 0 rather than a negative
value.

Besides, we also rewrite the code for randomly changing font size
and y position to prevent too much branch executions occurs in the
for loop.